+++ src/main/java/xyz/nikitacartes/easyauth/mixin/CommandManagerMixin.java
@@ -1,5 +1,6 @@
 package xyz.nikitacartes.easyauth.mixin;
 
+import com.mojang.brigadier.ParseResults;
 import net.minecraft.server.command.CommandManager;
 import net.minecraft.server.command.ServerCommandSource;
 import net.minecraft.util.ActionResult;
@@ -11,9 +12,9 @@ import xyz.nikitacartes.easyauth.event.AuthEventHandler;
 
 @Mixin(CommandManager.class)
 public class CommandManagerMixin {
-    @Inject(method = "execute(Lnet/minecraft/server/command/ServerCommandSource;Ljava/lang/String;)I", at = @At("HEAD"), cancellable = true)
-    private void checkCanUseCommands(ServerCommandSource source, String command, CallbackInfoReturnable<Integer> cir) {
-        ActionResult result = AuthEventHandler.onPlayerCommand(source.getPlayer(), command);
+    @Inject(method = "execute(Lcom/mojang/brigadier/ParseResults;Ljava/lang/String;)I", at = @At("HEAD"), cancellable = true)
+    private void checkCanUseCommands(ParseResults<ServerCommandSource> parseResults, String command, CallbackInfoReturnable<Integer> cir) {
+        ActionResult result = AuthEventHandler.onPlayerCommand(parseResults.getContext().getSource().getPlayer(), command);
         if (result == ActionResult.FAIL) {
             cir.setReturnValue(1);
         }