+++ src/main/java/xyz/nikitacartes/easyauth/mixin/ServerPlayNetworkHandlerMixin.java
@@ -23,15 +23,15 @@ public abstract class ServerPlayNetworkHandlerMixin {
     public ServerPlayerEntity player;
 
     @Inject(
-            method = "handleMessage(Lnet/minecraft/network/packet/c2s/play/ChatMessageC2SPacket;Lnet/minecraft/server/filter/FilteredMessage;)V",
+            method = "onChatMessage(Lnet/minecraft/network/packet/c2s/play/ChatMessageC2SPacket;)V",
             at = @At(
                     value = "INVOKE",
-                    target = "Lnet/minecraft/server/network/ServerPlayNetworkHandler;checkChatEnabled()Z",
+                    target = "Lnet/minecraft/server/network/ServerPlayNetworkHandler;canAcceptMessage(Ljava/lang/String;Ljava/time/Instant;Lnet/minecraft/network/message/LastSeenMessageList$Acknowledgment;)Z",
                     shift = At.Shift.AFTER
             ),
             cancellable = true
     )
-    private void onPlayerChat(ChatMessageC2SPacket chatMessageC2SPacket, FilteredMessage<String> message, CallbackInfo ci) {
+    private void onPlayerChat(ChatMessageC2SPacket packet, CallbackInfo ci) {
         ActionResult result = AuthEventHandler.onPlayerChat(this.player);
         if (result == ActionResult.FAIL) {
             ci.cancel();