+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 3) Fix crash if Fake Players join
 4) Fix softlock while log in inside a portal
 5) Add support for Floodgate players, thanks to @dmunozv04
+6) Probably fix missing server side translation [#36](https://github.com/NikitaCartes/EasyAuth/issues/36) with updating server translation
+7) Fix using command before login [#52](https://github.com/NikitaCartes/EasyAuth/issues/52), thanks to @were491
 
 ### 2.2.2, Minecraft 1.17, 1.17.1, 1.18[.1], 1.18.2