+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### 3.0.0 Minecraft 1.17, 1.17.1, 1.18[.1], 1.18.2, Minecraft 1.19
+### 3.0.0 Minecraft 1.17, 1.17.1, 1.18[.1], 1.18.2, 1.19, 1.19.1
 
 1) Fix MySQL support
 2) Add placeholder api support, thanks to @Nikijaz